Introduce interface for 'SharedItemsRepository'
